Migrate to CMake
Why migrate? CMake supports and generates "make files" for many supported compilers at this point, supports many platforms out of the box, and can easily be extended with additional features. In some cases, you can even have CMake download and install the necessary compilers for the project.

I've also split the teensy_loader_cli.c file into multiple parts, which aids in platform separation. While using #ifdef/#endif is fine mostly, there are better options available that do not require regenerating intermediate source code files. For now I've opted against removing all the #ifdef/#endif pairs, as some of them are still useful.
