+++ .gitignore
@@ -1,3 +1,1 @@
-teensy_loader_cli
-teensy_loader_cli.exe*
-*.swp
+build
